requirements.txt

pillow~=10.2.0
torchvision~=0.17.2
matplotlib~=3.8.0
seaborn~=0.12.2
lightning~=2.2.1
numpy~=1.26.4
yaml~=0.2.5
pyyaml~=6.0.1
munch~=2.5.0
torchmetrics~=1.3.1
torch~=2.2.2
codecarbon~=2.3.5
albumentations~=1.3.1
pybboxes~=0.1.6